Add a declaration for a multicore pipeline.
This adds handling for a new environment variable, CI_TARGET, which can
either be "mainline" (the default) or "multicore". With the latter, we
use Ocaml 4.10+multicore instead of the mainline compilers.

The value for opam_repository_commits moves to Conf, and now includes
ocaml-multicore/multicore-opam if CI_TARGET=multicore.

This also switches Conf.make_release to use OV.without_patch instead of
OV.with_just_major_and_minor, because the latter will throw away the
"+multicore" part (which should have no impact on the mainline pipeline).

Signed-off-by: Ewan Mellor <ewan@tarides.com>