Extend opam_build to cover packages that need to use make or other scripts rather than dune to perform their build.
This applies to batteries-included at the moment, which fails to build
using "dune build @install", but succeeds with "make all".  It also
applies to Coq and CompCert which use configure and make.

The new build strategies still use the dependencies from the solver, and
uses opam to install those, but uses then uses different commands to
execute the build.  There is a `Make strategy that simply calls make
with the specified targets, and there is a `Script strategy which allows
arbitrary shell commands to be passed in.

This adds a new stanza to conf.ml that specifies which packages this
applies to, and processes that in pipeline.ml.  This is not particularly
satisfactory, because it will only work with no more than one package
that needs to use a non-dune build. It falls back to the old behavior
otherwise (which presumably would fail).

Signed-off-by: Ewan Mellor <ewan@tarides.com>