Add the ability for ocaml-ci-local to run from multiple repos in parallel.
This adds Pipeline.local_test_multiple, and passes a label through
the analyze and build_with_docker jobs so that we can label the
node / cluster on the diagram depending on which repo we're working on.

This depends upon ewanmellor/ocurrent's expose-git-local-repo, so it
will need rebasing once that is merged for a PR to upstream.

Also, set docroot for Current_web.routes, so that we can serve the
static content for the website (the Ocaml logo).  This depends upon
ewanmellor/ocurrent's ui-restyle, so this will also need rebasing.

Signed-off-by: Ewan Mellor <ewan@tarides.com>