Extend the analyse job to handle non-opam-based repositories.
If there are no .opam files present in the repo, instead of failing the
job, return a placeholder analysis.  This contains the variants for
the build (derived from the given platforms in the same way as for Opam
builds) and the name of the package (derived from the source git repo).

This means that we can build CompCert, which does not have opam files
but we would still like to build in the presence of the ocaml compiler.

This change is not particularly satisfactory, and I'm not proposing it
for upstreaming yet.  I do not like the way that the platforms and
package name are passed in to the analyzer only for them to be passed
out again (as variants and package name).  However, the mechanism for
processing these things is all bundled together at the moment so it's
not easy to disentangle.

Signed-off-by: Ewan Mellor <ewan@tarides.com>