Add the ability to pull a fixed set of repos for inclusion in the test pipeline (along with those that we get as registered apps).
This refactors build, cluster_build, and dune_build_cache so that they
do not depend upon Current_github.  They were only using this to derive
a cache key, so we can just pass in a string.

Also, refactor pipeline heavily.

Allow the ocaml-multicore-ci-local task's dir parameter (the list of
repos to test) to be empty.  In that case, the fixed list of repos will
be used.

Allow 0-9 in the opam-repository sanitized local path.

Signed-off-by: Ewan Mellor <ewan@tarides.com>