Add a build_from_clone_component to the pipeline.
This adds an extra box to the build diagram for each of the builds, which
makes it much clearer.  It also means that we've got a caching point.

Pass the same label through analyse_build_summarise to
build_with_docker, to label the boxes in the diagram.
Pass the label from analyse_build_summarise to Analyse.
Change the formatting on that label.

Signed-off-by: Ewan Mellor <ewan@tarides.com>