Implement build suites that start from a specific compiler commit.
For a given commit in the compiler, we build the compiler itself
(into a switch named after the cset) and then build all our configured
packages using that compiler.

This adds an is_compiler flag to analyse, to change the behavior there:
we compile the compiler in the standard environment (i.e. without the
multicore-opam repo and not in the variant Docker images) and we skip
the solver phase, since the ocaml-variants.dev solve does not work in
this context.

It also adds a compiler_commit flag that gets passed through all the
build functionality, so that the generation of the ultimate build
Dockerfile knows which compiler it is supposed to be using.  This
creates a switch with that name, and then builds the compiler into it.
When checking out the build source, we now also check out the compiler
source (into /compiler-src in the Docker container).

This rearranges the build rules in opam_build a little, because we
need to insert the compiler rules early in the game.  Also, add some
comments to the Dockerfile to make it easier to see from the logs
what the intent is.

Kick all of this off by adding ocaml-multicore/ocaml-multicore@4.12+domains
and ocaml-multicore/ocaml-multicore@4.12+domains+effects to the static
configuration.