Extend the analysis phase to be able to analyse package builds in the context of a specified compiler build.
This changes our assumptions about which platform we are building for
(we don't build for all variants) and how we handle opam packages (we
don't expect to be able to pin them).

Change the build rules in this case so that the compiler is the sole source
repo in the build context (partially reverting changes in 8bb6c1bb5) and
we use "opam install" to build the packages under test.  This means that
builds will be compatible with ocluster (which can't have more than one
git repo in the build context).

Signed-off-by: Ewan Mellor <ewan@tarides.com>