Add irmin@2.7 to the list of fixed repos.
As part of this, drop all "vendor" opam files from the spec_opam_install
package list.  Irmin has vendored two tezos-context-hash packages that
are not in mainline opam-repository, and so they don't install.  Hack
around this by simply dropping them, though I'm not convinced that this
is a good solution in the long run.

Signed-off-by: Ewan Mellor <ewan@tarides.com>