Extend Dockerfile.web to support the new React-based UI.
This runs yarn build, copies the result into /srv in the container,
and configures the ocaml-multicore-ci-web command line to include
the --docroot and --admin-service-uri parameters.  The docroot command
line option is new, for this purpose.

Update the pin for opam-repository in both Dockerfiles, because the
menhir upstream archive has changed and the old pin no longer works.

This also adds ca-certificates to the web container, since something
has brought in mirage/ca-certs.

Signed-off-by: Ewan Mellor <ewan@tarides.com>