Extend the results reporting to include the compiler that we're using for the build.
This is essential if the builds are going to be understood correctly,
because previously we were just reporting the variant from the Docker
container that the build was performed in, but that's pretty useless if
the compiler didn't come from that container but was built fresh.

As part of this, remove the Current_github.Repo_id from Index.record's
interface (not all our repo references have come via the GitHub API) and
switch it to taking owner/name instead.  Change the interface so that
jobs are recorded one at a time, so that we don't have some variant builds
replacing the results of different variant builds.

Extend the pipeline to record results of cascaded builds.  When indexing
these builds, record them as e.g. "4.12+domains on debian-4.10" (and
add to pipeline_utils to construct that).

Signed-off-by: Ewan Mellor <ewan@tarides.com>