Different build_mechanism with the configurations
With the old version, the pipeline config file is little bit generalize,
but build_mechanism wasn't generalize.

Fix also repo url of sandmark's packages extractions

One thing to notice, the build_mechanism in the config file won't to be
taken when the analyze stage of the pipleline results on at least one
Selection.