Add rev-parse Sandmark pluggin
This plugin is used to get a hash(git rev-parse) of specific refs.

This is usefull because the clone function from ocurrent package
use already a prefix "origin/", the function fail when called with
a refs that is not prefixed with "origin/" by the git repo.

With that hash it's easly possible to fetch the right version.