Skip the analysis building Sandmark dependencies
Sandmark's package dependencies doesn't necessary need the all packages
in a repo. The actual state of multicore-ci includes all packages of
a given repo during the analysis. To avoid an unnecessary fail of the
pipeline we skip the analysis part.

Almost all those packages that sandmark use are old version and we fail
sometime during the analysis because of ocamlformat file when 'version'
is not specified.