Fix graphsql-ci failure after web-service restart
After the restart of the web-service(service/main.exe), any new
graphsql request fails, because when web-ui is starting, it uses
the value of the capability to reach the web-service instead of a reference,
and it ends up using the same value until we restart it.

This PR solves this issue by ensuring using the new capabality value,
after web-ui service reconnect to the web-service.